Feat: Added Error http status code handler and started getRelatedArticles service - partial
