+++ 01_presentation/exceptions/ResourceNotFound.js
@@ -1,0 +1,9 @@
+class ResourceNotFound extends Error {
+  constructor(msg) {
+    super()
+    this.message = msg
+    this.statusCode = 404
+  }
+}
+
+module.exports = ResourceNotFound
