+++ 02_domain/services/articleService.js
@@ -1,112 +1,6 @@
 const ServiceLocator = require('../../03_infra/dependencyInjection/serviceLocator')
 
 class ArticleService {
-  /* async get(page = 1, limit = 10, query = '', theme = '', category = '', author = '') {
-    // Para consultas diretas da home page
-    const homeQuery = req.query.home || false
-
-    // Para consultas que incluam artigos impulsionados na query principal
-    const boosted = req.query.boosted || false
-
-    // Procura por temas
-    const themes = {
-      $or: [
-        { 'theme.name': { $regex: `${theme}`, $options: 'i' } },
-        { 'theme.alias': { $regex: `${theme}`, $options: 'i' } }
-      ]
-    }
-
-    // Procura por categorias
-    const categories = {
-      $or: [
-        { 'category.name': { $regex: `${category}`, $options: 'i' } },
-        { 'category.alias': { $regex: `${category}`, $options: 'i' } }
-      ]
-    }
-
-    // Procura por autores
-    const authors = {
-      'author.name': { $regex: `${author}`, $options: 'i' }
-    }
-
-    // Procura por artigos publicados ou impulsionados
-    const publish = {
-      $or: [
-        { published: true },
-        { boosted: true }
-      ]
-    }
-
-    const onlyBoosted = {
-      boosted: true
-    }
-
-    // Procura por somente artigos não excluídos
-    const config = {
-      deleted: false
-    }
-
-    if (limit > 100) limit = 10
-
-    const boostedArticles = homeQuery ? await this.getBoostedArticles() : null
-
-    let count = await Article.aggregate([
-      {
-        $match: {
-          $and: [
-            {
-              $or: [
-                { title: { $regex: `${query}`, $options: 'i' } },
-                { shortDescription: { $regex: `${query}`, $options: 'i' } },
-                { longDescription: { $regex: `${query}`, $options: 'i' } }
-
-              ]
-            },
-            {
-              $or: [
-                themes,
-                categories
-              ]
-            },
-            { inactivated: false },
-            { deleted: false },
-            authors,
-            boosted ? onlyBoosted : publish,
-            config
-          ]
-        }
-      }]).count('id')
-
-    count = count.length > 0 ? count.reduce(item => item).id : 0
-
-    Article.aggregate([
-      {
-        $match: {
-          $and: [
-            {
-              $or: [
-                { title: { $regex: `${query}`, $options: 'i' } },
-                { shortDescription: { $regex: `${query}`, $options: 'i' } },
-                { longDescription: { $regex: `${query}`, $options: 'i' } }
-              ]
-            },
-            {
-              $or: [
-                themes,
-                categories
-              ]
-            },
-            { inactivated: false },
-            { deleted: false },
-            authors,
-            boosted ? onlyBoosted : publish,
-            config
-          ]
-        }
-      }, { $sort: { publishAt: -1 } }])
-      .skip(page * limit - limit).limit(limit).then(articles => res.json({ articles, count, limit, boostedArticles }))
-  } */
-
   async getBoostedArticles(skip = 0, take = 5) {
     return ServiceLocator.unitOfWork.articleRepository.getBoosted(skip, take)
   }
@@ -115,42 +9,13 @@ class ArticleService {
     return ServiceLocator.unitOfWork.articleRepository.getByCustomUri(customUri, ['boosted', 'published'])
   }
 
-  /* async getRelateds(customURL, limit = 3) {
-    try {
-      if (limit > 10) limit = 3
-      if (!customURL) {
-        let articles = await getBoostedArticles()
+  async getPopularArticles(skip = 0, take = 10) {
+    return ServiceLocator.unitOfWork.articleRepository.getPopulars(skip, take)
+  }
 
-        articles = articles.articles.filter((elem, index) => index <= (limit - 1))
-        return res.json(articles)
-      } else {
-        const article = await Article.findOne({ customURL })
-
-        if (!article) {
-          let articles = await getBoostedArticles()
-          articles = articles.articles.filter((elem, index) => index <= (limit - 1))
-
-          return res.json(articles)
-        }
-
-        await Article.aggregate([
-          {
-            $match: {
-              $or: [
-                { 'theme._id': article.theme._id },
-                { 'category._id': article.category._id }
-              ],
-              _id: { $ne: article._id },
-              published: true,
-              inactivated: false,
-              deleted: false
-            }
-          }]).limit(limit).then(response => res.json(response))
-      }
-    } catch (error) {
-      return res.status(500).send('Ocorreu um erro interno ao obter as informações, tente novamente mais tarde')
-    }
-  } */
+  async getRelateds(articleUri, limit = 5) {
+    return ServiceLocator.unitOfWork.articleRepository.getRelateds(articleUri, limit)
+  }
 }
 
 module.exports = new ArticleService()
