+++ 03_infra/services/index.js
@@ -2,6 +2,7 @@ const DI = require('../dependencyInjection')
 const Database = require('../database')
 const Middlewares = require('../api/middlewares')
 const Actions = require('../../01_presentation/actions/actions')
+const HttpStatusCodeHandler = require('../../01_presentation/handlers/httpstatuscode')
 
 const Cors = require('../api/cors')
 
@@ -20,7 +21,7 @@ class Services {
   }
 
   static configureActions(express) {
-
+    HttpStatusCodeHandler.Configure(express)
   }
 
   static start(express, port = 3010, host = '0.0.0.0') {
